techpack: display: Only skip color inversion filter
So far only color inversion filter has been problematic with our fingerprint dimming implementation as it brightens the screen instead of dimming it when using that filter.
Sometimes the filter does not get skipped at the correct time causing our users to think that their screen has bugged out so we only skip the color inversion filter for now as skipping that works as intended.

Signed-off-by: Abhay Singh Gill <abhaygill017@gmail.com>